Fixed @smarterclayton's comments from #384 part 2

+++ pkg/deploy/controller/deployment_controller.go
@@ -136,17 +136,23 @@ func (dc *DeploymentController) HandlePod() {
 	switch pod.CurrentState.Status {
 	case kapi.PodRunning:
 		nextDeploymentStatus = deployapi.DeploymentStatusRunning
-	case kapi.PodSucceeded:
+	case kapi.PodSucceeded, kapi.PodFailed:
 		nextDeploymentStatus = deployapi.DeploymentStatusComplete
-		// Automatically clean up successful pods
-		if err := dc.PodControl.deletePod(deployment.Namespace, pod.Name); err != nil {
-			glog.V(4).Infof("Couldn't delete completed pod %s for deployment %s: %#v", pod.Name, deployment.Name, err)
-		} else {
-			glog.V(4).Infof("Deleted completed pod %s for deployment %s", pod.Name, deployment.Name)
-		}
-	case kapi.PodFailed:
 		// Detect failure based on the container state
-		nextDeploymentStatus = deployapi.DeploymentStatusFailed
+		for _, info := range pod.CurrentState.Info {
+			if info.State.Termination != nil && info.State.Termination.ExitCode != 0 {
+				nextDeploymentStatus = deployapi.DeploymentStatusFailed
+			}
+		}
+
+		// Automatically clean up successful pods
+		if nextDeploymentStatus == deployapi.DeploymentStatusComplete {
+			if err := dc.PodControl.deletePod(deployment.Namespace, pod.Name); err != nil {
+				glog.V(4).Infof("Couldn't delete completed pod %s for deployment %s: %#v", pod.Name, deployment.Name, err)
+			} else {
+				glog.V(4).Infof("Deleted completed pod %s for deployment %s", pod.Name, deployment.Name)
+			}
+		}
 	}
 
 	if deployment.Status != nextDeploymentStatus {
